feat: support setting running container as dev container

+++ cmd/agent/workspace/delete.go
@@ -82,11 +82,15 @@ func removeContainer(ctx context.Context, workspaceInfo *provider2.AgentWorkspac
 		return err
 	}
 
-	err = runner.Delete(ctx)
-	if err != nil {
-		return err
+	if workspaceInfo.Workspace.Source.Container != "" {
+		log.Infof("Skipping container deletion, since it was not created by DevPod")
+	} else {
+		err = runner.Delete(ctx)
+		if err != nil {
+			return err
+		}
+		log.Debugf("Successfully removed DevPod container from server")
 	}
-	log.Debugf("Successfully removed DevPod container from server")
 
 	return nil
 }
